perf(fonti): riduce caricamento e pacchetto dello stato fonti (#394)
