+++ src/lib/integrated-sources.ts
@@ -1,4 +1,6 @@
 import "server-only";
+import { getIntegratedLoadState, IntegratedLoadOverloadedError, withIntegratedDatasetLoadSlot } from "@/lib/integrated-load-limiter";
+export { IntegratedLoadOverloadedError, withIntegratedDatasetLoadSlot } from "@/lib/integrated-load-limiter";
 
 import { constants } from "node:fs";
 import { open } from "node:fs/promises";
@@ -47,17 +49,8 @@ const MAX_SOURCE_CATALOG_BYTES = 16 * 1024 * 1024;
 const MAX_DATASET_PROOF_BYTES = 1024 * 1024;
 const MAX_DATASET_CATALOG_BYTES = 1024 * 1024;
 const MAX_COMPRESSED_DATASET_CHUNK_BYTES = INTEGRATED_ROW_CHUNK_MAX_RAW_BYTES + 64 * 1024;
-const MAX_CONCURRENT_DATASET_LOADS = 2;
-const MAX_PENDING_DATASET_LOADS = 64;
 const MAX_CONSUMERS_PER_CHUNK = 64;
 const BOUNDED_READ_BLOCK_BYTES = 64 * 1024;
-
-export class IntegratedLoadOverloadedError extends Error {
-  constructor() {
-    super("Il caricamento dei dati integrati è temporaneamente saturo.");
-    this.name = "IntegratedLoadOverloadedError";
-  }
-}
 
 export type IntegratedSourceBundle = {
   release: IntegratedReleaseProof;
@@ -83,16 +76,7 @@ type InFlightChunkLoad = {
   settled: boolean;
 };
 
-type QueuedDatasetLoad = {
-  resolve: () => void;
-  reject: (reason: unknown) => void;
-  signal: AbortSignal;
-  onAbort: () => void;
-};
-
 const inFlightChunkLoads = new Map<string, InFlightChunkLoad>();
-const datasetLoadQueue: QueuedDatasetLoad[] = [];
-let activeDatasetLoads = 0;
 let completedChunkLoads = 0;
 let maxObservedChunkRawBytes = 0;
 
@@ -104,56 +88,6 @@ function abortedError(): Error {
 
 function throwIfAborted(signal?: AbortSignal): void {
   if (signal?.aborted) throw abortedError();
-}
-
-function acquireDatasetLoadSlot(signal: AbortSignal): Promise<void> {
-  throwIfAborted(signal);
-  if (activeDatasetLoads < MAX_CONCURRENT_DATASET_LOADS) {
-    activeDatasetLoads += 1;
-    return Promise.resolve();
-  }
-  if (datasetLoadQueue.length >= MAX_PENDING_DATASET_LOADS) {
-    return Promise.reject(new IntegratedLoadOverloadedError());
-  }
-  return new Promise((resolve, reject) => {
-    const queued = {} as QueuedDatasetLoad;
-    const onAbort = () => {
-      const index = datasetLoadQueue.indexOf(queued);
-      if (index >= 0) datasetLoadQueue.splice(index, 1);
-      reject(abortedError());
-    };
-    Object.assign(queued, { resolve, reject, signal, onAbort });
-    signal.addEventListener("abort", onAbort, { once: true });
-    datasetLoadQueue.push(queued);
-  });
-}
-
-function releaseDatasetLoadSlot(): void {
-  activeDatasetLoads -= 1;
-  while (datasetLoadQueue.length > 0) {
-    const next = datasetLoadQueue.shift()!;
-    next.signal.removeEventListener("abort", next.onAbort);
-    if (next.signal.aborted) {
-      next.reject(abortedError());
-      continue;
-    }
-    activeDatasetLoads += 1;
-    next.resolve();
-    break;
-  }
-}
-
-export async function withIntegratedDatasetLoadSlot<T>(
-  load: () => Promise<T>,
-  signal: AbortSignal,
-): Promise<T> {
-  await acquireDatasetLoadSlot(signal);
-  try {
-    throwIfAborted(signal);
-    return await load();
-  } finally {
-    releaseDatasetLoadSlot();
-  }
 }
 
 async function readRegular(
@@ -472,11 +406,8 @@ function attachChunkConsumer(
 /** Test-only observability for the bounded, server-internal artifact loader. */
 export function getIntegratedDatasetLoaderDiagnosticsForTests() {
   return {
-    maxConcurrentLoads: MAX_CONCURRENT_DATASET_LOADS,
-    maxPendingLoads: MAX_PENDING_DATASET_LOADS,
+    ...getIntegratedLoadState(),
     maxConsumersPerChunk: MAX_CONSUMERS_PER_CHUNK,
-    activeLoads: activeDatasetLoads,
-    queuedLoads: datasetLoadQueue.length,
     inFlightChunkKeys: [...inFlightChunkLoads.keys()],
     completedChunkLoads,
     maxObservedChunkRawBytes,
@@ -484,7 +415,8 @@ export function getIntegratedDatasetLoaderDiagnosticsForTests() {
 }
 
 export function resetIntegratedDatasetLoaderDiagnosticsForTests(): void {
-  if (activeDatasetLoads !== 0 || datasetLoadQueue.length !== 0 || inFlightChunkLoads.size !== 0) {
+  const { activeLoads, queuedLoads } = getIntegratedLoadState();
+  if (activeLoads !== 0 || queuedLoads !== 0 || inFlightChunkLoads.size !== 0) {
     throw new Error("Impossibile azzerare la diagnostica durante un caricamento.");
   }
   completedChunkLoads = 0;
